pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-similarity-engine"
version = "0.2.7"
description = ""
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.9"
authors = [
    {name = "Jonathan Louis"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["code", "analysis", "refactoring", "cli"]
dependencies = [
    "click>=8.0",
    "numpy>=1.21",
    "scikit-learn>=1.0",
    "tree-sitter>=0.21",
    "tree-sitter-python>=0.21",
    "tree-sitter-swift>=0.0.1",
    "tree-sitter-rust>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-go>=0.21",
    "tree-sitter-c>=0.21",
    "tree-sitter-cpp>=0.22",
    "tree-sitter-java>=0.21",
    "tree-sitter-c-sharp>=0.21",
    "tree-sitter-php>=0.22",
    "tree-sitter-ruby>=0.21",
    "tree-sitter-kotlin>=0.3",
    "tree-sitter-scala>=0.21",
    "tree-sitter-lua>=0.1",
    "tree-sitter-bash>=0.21",
    "llama-cpp-python>=0.2.0",
    "huggingface-hub>=0.20.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
cse = "code_similarity_engine.cli:cli"
cse-download-models = "code_similarity_engine.model_manager:download_models_cli"

[project.urls]
Repository = "https://github.com/jonathanlouis/code-similarity-engine"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
